synth_parser: Fix volume of extended notes
Extended notes (e.g., a2^4, or a LA that last 3/4) are notes that are
synthesized as various notes of (possibly) varying durations. Since a
single note ends up becoming multiple notes, the envelope of the original
note cannot be used as is in the various "sub-notes". If it were used,
each "sub-note" would play independently from each other, instead of
having the same envelope wrap the entire extended note.

Create a custom envelope for each sub-note in an extended note, to properly
expand the envelope of the original extended note among the sub-notes.